refactor: hide helpers behind internal package
